+++ src/test/java/config/FileReaderTest.java
@@ -1,4 +1,4 @@
-package policy;
+package config;
 
 import domain.Attendance;
 import domain.AttendanceSheet;
@@ -11,27 +11,27 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-public class FileReaderPolicyTest {
+public class FileReaderTest {
     private static final String FILE_PATH = "src/main/resources/attendances.csv";
 
-    FileReaderPolicy fileReaderPolicy;
+    FileReader fileReader;
 
     @BeforeEach
     void setUp() {
-        fileReaderPolicy = new FileReaderPolicy(FILE_PATH);
+        fileReader = new FileReader(FILE_PATH);
     }
 
     @Test
     @DisplayName("지정된 위치의 파일이 아니면 예외가 발생한다")
     public void validateFileReaderPolicyTest() {
-        assertThatThrownBy(() -> new FileReaderPolicy("attendances.csv"))
+        assertThatThrownBy(() -> new FileReader("attendances.csv"))
                 .isInstanceOf(RuntimeException.class);
     }
 
     @Test
     @DisplayName("파일을 읽어들일 수 있다")
     public void readLinesTest() {
-        assertThat(fileReaderPolicy.readLines())
+        assertThat(fileReader.readLines())
                 .isInstanceOf(List.class);
     }
 
@@ -42,7 +42,7 @@ public class FileReaderPolicyTest {
         String[] splitLine = {"링크"};
 
         //when-then
-        assertThatThrownBy(() -> fileReaderPolicy.validateSplitLineFormat(splitLine))
+        assertThatThrownBy(() -> fileReader.validateSplitLineFormat(splitLine))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -53,7 +53,7 @@ public class FileReaderPolicyTest {
         String attendanceDateTime = "2024:12:13 09:11";
 
         //when-then
-        assertThatThrownBy(() -> fileReaderPolicy.parseAttendanceDateTime(attendanceDateTime))
+        assertThatThrownBy(() -> fileReader.parseAttendanceDateTime(attendanceDateTime))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -64,7 +64,7 @@ public class FileReaderPolicyTest {
         String attendanceInfo = "쿠키,2024-12-13 10:08";
 
         //when-then
-        assertThat(fileReaderPolicy.createAttendance(attendanceInfo))
+        assertThat(fileReader.createAttendance(attendanceInfo))
                 .isInstanceOf(Attendance.class);
     }
 
@@ -78,7 +78,7 @@ public class FileReaderPolicyTest {
                 );
 
         //when-then
-        assertThat(fileReaderPolicy.createAttendances(attendancesInfo))
+        assertThat(fileReader.createAttendances(attendancesInfo))
                 .isInstanceOf(AttendanceSheet.class);
     }
